pyproject.toml
[project]
name = "lingchu-bot"
version = "0.0.0"
description = "lingchu-bot"
readme = "README.md"
requires-python = ">=3.13.0, <4.0.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-sentry>=2.0.0",
    "nonebot-plugin-orm[aiosqlite]>=0.8.2",
    "annotated-doc>=0.0.4",
    "binaryornot>=0.4.4",
    "black>=25.11.0",
    "bracex>=2.6",
    "bump-my-version>=1.2.4",
    "cashews>=7.4.3",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.4",
    "cookiecutter>=2.6.0",
    "deepmerge>=2.0",
    "distlib>=0.4.0",
    "filelock>=3.20.0",
    "httptools>=0.7.1",
    "idna>=3.11",
    "isort>=7.0.0",
    "linkify-it-py>=2.0.3",
    "markdown>=3.10",
    "markdown-it-py[linkify]>=4.0.0",
    "mdit-py-plugins>=0.5.0",
    "mdurl>=0.1.2",
    "mypy-extensions>=1.1.0",
    "noneprompt>=0.1.10",
    "nonestorage>=0.1.0",
    "pastel>=0.2.1",
    "pathspec>=0.12.1",
    "poethepoet>=0.38.0",
    "prompt-toolkit>=3.0.52",
    "propcache>=0.4.1",
    "pyfiglet>=1.0.4",
    "pygments>=2.19.2",
    "pygtrie>=2.5.0",
    "pymdown-extensions>=10.17.1",
    "pyright[nodejs]>=1.1.407",
    "python-slugify>=8.0.4",
    "pytokens>=0.3.0",
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "requests>=2.32.5",
    "ruff>=0.14.6",
    "text-unidecode>=1.3",
    "textual>=6.6.0",
    "tomlkit>=0.13.3",
    "typing-inspection>=0.4.2",
    "uc-micro-py>=1.0.3",
    "urllib3>=2.5.0",
    "virtualenv>=20.35.4",
    "watchfiles>=1.1.1",
    "wcmatch>=10.1",
    "wcwidth>=0.2.14",
    "zensical>=0.0.9",
    "nicegui>=3.3.1",
]

[project.optional-dependencies]
dev = [
    "pyright[nodejs]",
    "ruff",
    "bump-my-version>=1.2.4",
    "poethepoet>=0.37.0",
    "black>=25.11.0",
    "isort>=7.0.0",
    "zensical>=0.0.9",
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
nonebot-plugin-localstore = ["nonebot_plugin_localstore"]
nonebot-plugin-sentry = ["nonebot_plugin_sentry"]
nonebot-plugin-orm = ["nonebot_plugin_orm"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
# For more rules, see https://docs.astral.sh/ruff/rules/.
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "C90",   # mccabe
    "N",     # pep8-naming
    "PL",    # pylint
    "UP",    # pyupgrade
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "BLE",   # flake8-blind-except
    "FBT",   # flake8-boolean-trap
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "ICN",   # flake8-import-conventions
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "SLOT",  # flake8-slots
    "TID",   # flake8-tidy-imports
    "TC",    # flake8-type-checking
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    # "ERA",   # eradicate
    "FAST", # FastAPI
    "PERF", # Perflint
    "PGH",  # pygrep-hooks
    "FURB", # refurb
    "TRY",  # tryceratops
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E402",  # module-import-not-at-top-of-file  # nonebot2 require() violates this
    "B008",  # function-call-in-default-argument  # nonebot2 Depends() without Annotated violates this
    "UP037", # quoted-annotation
    # "RUF001",  # ambiguous-unicode-character-string
    # "RUF002",  # ambiguous-unicode-character-docstring
    # "RUF003",  # ambiguous-unicode-character-comment
    # "ANN201",  # missing-return-type-undocumented-public-function
    "ANN202",  # missing-return-type-private-function
    "ANN401",  # any-type
    "COM812",  # missing-trailing-comma
    "PLC0415", # import-outside-top-level
    "TID252",  # prefer-absolute-imports
]
allowed-confusables = [
    "，",
    "。",
    "“",
    "”",
    "：",
    "；",
    "？",
    "！",
    "【",
    "】",
    "《",
    "》",
    "…",
    "—",
    "（",
    "）",
    "、",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
# For Pylance/Pyright configurations, see https://microsoft.github.io/pyright/#/configuration.
pythonVersion = "3.13"
pythonPlatform = "All"
typeCheckingMode = "standard"

[tool.pdm.scripts]
dev-app = "pdm run bot.py"
dev-docs = "zensical serve"
lint = "pdm run ruff check . --fix"
format = "pdm run isort . && pdm run black . && pdm run ruff format ."
build = "pdm run format && pdm run lint && pdm build"
check-all = "pdm run format && pdm run lint"
[tool.poe.tasks]
# lint = ["_lint-ruff"]
# _lint-ruff = "pdm run ruff check . --fix"
# format = ["_format-black", "_format-isort", "_format-ruff"]
# _format-black = "pdm run black ."
# _format-isort = "pdm run isort ."
# _format-ruff = "pdm run ruff format ."
# build = ["format", "lint", "_build"]
# _build = "pdm build"
